fix imports and rename models for schemes
