+++ routes/index.js
@@ -1,1 +1,2 @@
 exports.userRouter = require('./users');
+exports.authRouter = require('./auth');
